+++ src/structum/cli/__init__.py
@@ -1,0 +1,10 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: 2025 PythonWoods
+
+"""
+CLI Package.
+"""
+
+from .main import app
+
+__all__ = ["app"]