+++ src/structum/plugins/sample/__init__.py
@@ -1,0 +1,8 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: 2025 PythonWoods
+
+"""Sample Plugin Package."""
+
+from .plugin import SamplePlugin
+
+__all__ = ["SamplePlugin"]